catch upload img error

+++ src/api/services/product.service.js
@@ -11,7 +11,7 @@ module.exports = {
     getProductById: async (productId) => {
         try {
             if (!mongoose.Types.ObjectId.isValid(productId)) {
-                throw new createError(404, 'Product not found!');
+                throw new createError(404, 'Invalid input!');
             }
             const res = await Product.find({ _id: productId });
             console.log(res);
@@ -26,7 +26,11 @@ module.exports = {
     },
     CreateProduct: async (req) => {
         try {
-            const { name, price, category, quantity, description, rating } = req.body;
+            const { name, price, category, description, rating } = req.body;
+            if(!name || !description || !price) {
+                throw new CreateError(400, "Invalid input");
+            }
+            if(!req.file) throw new CreateError(400, "Please upload file!")
             const result = await cloudinary.uploader.upload(req.file.path);
             await deleteFile(req.file.path);
             let product = new Product({
@@ -35,7 +39,6 @@ module.exports = {
                 cloudinary_id: result.public_id,
                 price,
                 category,
-                quantity,
                 description,
                 rating,
             });
@@ -76,7 +79,7 @@ module.exports = {
     deleteProductById: async (id) => {
         try {
             if (!mongoose.Types.ObjectId.isValid(id)) {
-                throw new createError(404, 'Product not found');
+                throw new createError(400, 'Invalid input!');
             }
             const product = await Product.findById(id);
 
@@ -99,7 +102,7 @@ module.exports = {
     updateProductById: async ({ params, body, file }) => {
         try {
             
-            const { name, description, price, quantity, rating, category } = body;
+            const { name, description, price, rating, category } = body;
             if (!mongoose.Types.ObjectId.isValid(params.id)) {
                 throw new createError(404, 'Product not found');
             }
@@ -123,7 +126,6 @@ module.exports = {
                         image: result.secure_url,
                         cloudinary_id: result.public_id,
                         price,
-                        quantity,
                         rating,
                         category,
                     },
